Moved integration/auth_test.go to its own package. Moved fileKeySource to auth_test.go
