+++ integration/auth/auth_test.go
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package integration
+// Package auth contains integration tests for the firebase.google.com/go/auth package.
+package auth
 
 import (
 	"bytes"
